Don't use magic numbers for indexing arrays
We were using magic numbers to index the arrays for the score
calculation. These indices represented the 710nm and 900nm values for
the example data. I have changed the code to make this more explicit and
to allow it to handle data with indexing different to that of the
example data.